serial-read/requirements.txt
pyserial ==3.5
